+++ guests/eth-block/Cargo.toml
@@ -16,3 +16,7 @@ radium = { git = 'https://github.com/ferrilab/radium.git', rev = "723bed5" }
 crypto-bigint = { git = "https://github.com/risc0/RustCrypto-crypto-bigint", tag = "v0.5.2-risc0" }
 k256 = { git = "https://github.com/risc0/RustCrypto-elliptic-curves", tag = "k256/v0.13.1-risc0" }
 sha2 = { git = "https://github.com/risc0/RustCrypto-hashes", tag = "sha2/v0.10.6-risc0" }
+
+[profile.release]
+codegen-units = 1
+panic = "abort"
